# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Felix Doggy API
  description: |
    REST API for the Felix Doggy e-commerce platform — a dog-themed store
    selling dogs, dog food, toys, and clothes.

    ## Authentication
    Protected endpoints require a JWT Bearer token obtained via `POST /api/auth/login`
    or `POST /api/auth/verify-otp`. Pass it as `Authorization: Bearer <token>`.

    ## Authorization Roles
    | Role    | Description                                      |
    |---------|--------------------------------------------------|
    | `GUEST` | Unauthenticated visitor. Can browse and checkout. |
    | `USER`  | Registered, verified customer.                   |
    | `ADMIN` | Full administrative access.                      |

    ## Rate Limiting
    All endpoints are subject to a global rate limit (100 req / 10 s).
    Specific endpoints apply stricter limits as documented.
  version: 1.0.0
  contact:
    name: Felix Doggy Dev Team
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://record-store-api.onrender.com
    description: Production (Render)
...
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token issued on login or OTP verification.
        Payload contains `{ userId: string }`, expires in 7 days.

  # ─── Reusable Schemas ────────────────────────────────────────────────────────
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Abbey Road"
        artist:
          type: string
          example: "The Beatles"
        price:
          type: number
          format: float
          example: 29.99
        imgUrl:
          type: string
          format: uri
          example: "https://images.unsplash.com/photo-1590419690008-905895e8fe0d?w=600&h=600&fit=crop"
        category:
          type: string
          enum: [dogs, food, toys, clothes]
          example: "dogs"
        stock:
          type: integer
          minimum: 0
          example: 50
        description:
          type: [string, "null"]
          example: "Extremely talkative, does dramatic howling concerts at 3 AM."
      required: [id, title, artist, price, imgUrl, category, stock]
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        userId:
          type: [string, "null"]
          format: uuid
          description: "Null for guest checkouts."
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        customerEmail:
          type: [string, "null"]
          example: "buyer@example.com"
        customerPhone:
          type: [string, "null"]
          example: "0901234567"
        shippingAddr:
          type: [string, "null"]
          example: "456 Nguyễn Huệ, Q1, TP.HCM"
        totalAmount:
          type: number
          format: float
          example: 59.98
        status:
          type: string
          enum: [PENDING, PROCESSING, SHIPPED, COMPLETED, CANCELLED]
          default: PENDING
          example: "PENDING"
        createdAt:
          type: string
          format: date-time
      required: [id, totalAmount, status, createdAt]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        orderId:
          type: string
          format: uuid
        productId:
          type: integer
          example: 3
        quantity:
          type: integer
          minimum: 1
          example: 2
        priceAtTime:
          type: number
          format: float
          description: Snapshot of product price at time of purchase.
          example: 29.99
      required: [id, orderId, productId, quantity, priceAtTime]

    # ── Public Response DTOs ───────────────────────────────────────────────────
...
    OrderWithRelations:
      type: object
      description: Order with nested user and order items (including product details).
      allOf:
        - $ref: '#/components/schemas/Order'
        - type: object
          properties:
            user:
              type: [object, "null"]
              properties:
                fullName:
                  type: [string, "null"]
                email:
                  type: string
                  format: email
            orderItems:
              type: array
              items:
                $ref: '#/components/schemas/OrderItemWithProduct'
    OrderItemWithProduct:
      type: object
      allOf:
        - $ref: '#/components/schemas/OrderItem'
        - type: object
          properties:
            product:
              $ref: '#/components/schemas/Product'

    # ── Request DTOs ───────────────────────────────────────────────────────────
...
    CheckoutRequest:
      type: object
      properties:
        customerEmail:
          type: string
          format: email
          example: "buyer@example.com"
        customerPhone:
          type: string
          example: "0901234567"
        customerName:
          type: string
          description: Optional. If provided and user is authenticated with an empty fullName, it will be saved to their profile.
          example: "Trần Thị C"
        shippingAddr:
          type: string
          example: "789 Hai Bà Trưng, Q3, TP.HCM"
        items:
          type: array
          minItems: 1
          items:
            type: object
            properties:
              id:
                type: integer
                description: Product ID.
                example: 3
              quantity:
                type: integer
                minimum: 1
                example: 2
            required: [id, quantity]
      required: [customerEmail, customerPhone, shippingAddr, items]
...
    CheckoutResponse:
      type: object
      properties:
        message:
          type: string
          example: "Order created successfully"
        order:
          allOf:
            - $ref: '#/components/schemas/Order'
            - type: object
              properties:
                orderItems:
                  type: array
                  items:
                    $ref: '#/components/schemas/OrderItem'
      required: [message, order]
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Error description"
      required: [message]
    RateLimitResponse:
      type: object
      description: Returned when rate limit is exceeded (HTTP 429).
      properties:
        message:
          type: string
          example: "Bạn đã gửi quá nhiều yêu cầu. Vui lòng thử lại sau 10 giây."
      required: [message]

  # ─── Shared Parameters ──────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            noToken:
              summary: No token provided
              value:
                message: "No token provided"
            invalidToken:
              summary: Invalid or expired token
              value:
                message: "Invalid or expired token"
            userNotFound:
              summary: Token valid but user deleted
              value:
                message: "User not found"
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: "Internal Server Error"
    RateLimited:
      description: Rate limit exceeded (HTTP 429).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitResponse'

# ═══════════════════════════════════════════════════════════════════════════════
# PATHS
# ═══════════════════════════════════════════════════════════════════════════════
paths:
...
  /api/orders/checkout:
    post:
      tags: [Orders]
      summary: Checkout (create order)
      description: |
        Creates a new order. Supports both authenticated and guest checkouts.
        If an `Authorization` header is present, the order is linked to the
        user and empty profile fields may be auto-filled.
        Rate-limited by `strictLimiter` (10 req / 10 s).
        Runs inside a database transaction with retry (3 attempts).
      operationId: checkout
      security:
        - BearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '201':
          description: Order created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
        '400':
          description: Empty cart, product not found, or insufficient stock.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emptyCart:
                  summary: Cart is empty
                  value:
                    message: "Cart is empty"
                notFound:
                  summary: Product not found
                  value:
                    message: "Product ID 999 not found"
                noStock:
                  summary: Insufficient stock
                  value:
                    message: "Not enough stock for Abbey Road"
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/orders/my-orders:
    get:
      tags: [Orders]
      summary: Get my orders
      description: Returns all orders belonging to the authenticated user, most recent first.
      operationId: getMyOrders
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of user's orders with product details.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderWithRelations'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── Admin ────────────────────────────────────────────────────────────────────
...